Move functionality related to waiting on DOM into utils.js
